ingest/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy[asyncio]>=2.0.36
psycopg[binary]>=3.2.0
pgvector>=0.3.6
httpx>=0.28.0
watchdog>=6.0.0
pypdf>=5.0.0
python-docx>=1.1.0
pydantic-settings>=2.6.0
structlog>=24.4.0
